register and login

+++ src/main/java/com/paw/controller/TableController.java
@@ -11,14 +11,17 @@ import org.osgl.mvc.annotation.PostAction;
 import org.osgl.mvc.annotation.PutAction;
 import org.osgl.mvc.result.BadRequest;
 import org.osgl.mvc.result.NotFound;
+import org.osgl.mvc.result.Unauthorized;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import java.util.List;
+import java.util.stream.Collectors;
+
 
 @Controller
-public class TableController {
+public class TableController extends AuthenticatedController{
 
     private static final Logger LOG = LoggerFactory.getLogger(TableController.class);
 
@@ -30,6 +33,7 @@ public class TableController {
 
     @PostAction("/table")
     public PawTable addTable(PawTable pawTable) {
+        pawTable.setOwner(me.email);
         LOG.info("Dodano tablice: " + pawTable.toString());
         return tableDao.save(pawTable);
 
@@ -37,16 +41,27 @@ public class TableController {
 
     @GetAction("/table")
     public List<PawTable> getTableList() {
-        try {
-            return tableDao.findAllAsList();
-        } catch (NullPointerException e) {
-            throw new NotFound();
+        if(me !=null) {
+            try {
+                return tableDao.findAllAsList().stream()
+                        .filter(table -> table.getOwner().equals(me.email()))
+                        .collect(Collectors.toList());
+            } catch (NullPointerException e) {
+                throw new NotFound();
+            }
+        } else {
+            throw new Unauthorized();
         }
     }
 
     @GetAction("/table/{id}")
     public PawTable getTableDetails(int id) {
-        return tableDao.findById(id);
+        PawTable table = tableDao.findById(id);
+        if(table.getOwner().equals(me.email())) {
+            return table;
+        } else {
+            throw new Unauthorized();
+        }
     }
 
     @PutAction("/table/{id}")
@@ -55,30 +70,48 @@ public class TableController {
             throw new BadRequest("Id in body do not match id in path");
         } else {
             PawTable tempTable = tableDao.findById(id);
-            LOG.info("Zmieniono nazwe tablicy o  id:{} z {} na {}", id, tempTable.getName(), pawTable.getName());
-            tempTable.setName(pawTable.getName());
-            return tableDao.save(tempTable);
+            if(tempTable.getOwner().equals(me.email())) {
+                LOG.info("Zmieniono nazwe tablicy o  id:{} z {} na {}", id, tempTable.getName(), pawTable.getName());
+                tempTable.setName(pawTable.getName());
+                return tableDao.save(tempTable);
+            } else {
+                throw new Unauthorized();
+            }
         }
     }
 
     @DeleteAction("/table/{id}")
     public void deleteTableById(int id) {
-        tableDao.deleteById(id);
+        PawTable tempTable = tableDao.findById(id);
+        if(tempTable.getOwner().equals(me.email())) {
+            tableDao.deleteById(id);
+        } else {
+            throw new Unauthorized();
+        }
     }
 
     @PostAction("table/{id}/list")
     public PawTable addNewList(int id, PawList pawList) {
         PawTable tempTable = tableDao.findById(id);
-        LOG.info("Dodano liste {} do tablicy o id : {}", pawList.toString(), tempTable.getId());
-        listDao.save(pawList);
-        tempTable.addPawList(pawList);
-        return tableDao.save(tempTable);
+        if(tempTable.getOwner().equals(me.email())) {
+            LOG.info("Dodano liste {} do tablicy o id : {}", pawList.toString(), tempTable.getId());
+            listDao.save(pawList);
+            tempTable.addPawList(pawList);
+            return tableDao.save(tempTable);
+        } else {
+            throw new Unauthorized();
+        }
     }
 
 
     @GetAction("table/{id}/list")
     public List<PawList> getPawListListOnTable(int id) {
-        return tableDao.findById(id).getPawLists();
+        PawTable table = tableDao.findById(id);
+        if(table.getOwner().equals(me.email())) {
+            return table.getPawLists();
+        } else {
+            throw new Unauthorized();
+        }
     }
 
 }
